Refactored code to enhance exception messages and efficient handling of XBRL data
This patch considerably enhances the exception messages, improves logic of the managers for parsing XBRL data and makes "yield" use for efficient memory consumption. Also, modified decimal parse in ixbrl parser to handle cases of non-integer decimals. Lastly, updated tests according to these changes and removed unnecessary data cleanup operations in tests. Expected to improve code maintainability and execution efficiency overall.

+++ app/parser/ixbrl_parser.py
@@ -163,7 +163,7 @@ class IxbrlParser(BaseXBRLParser):
             context_category = context_parts[2] if len(context_parts) > 2 else None
 
             # _____attr[decimals]
-            decimals = int(tag.get("decimals")) if tag.get("decimals") else None
+            decimals = float(tag.get("decimals")) if tag.get("decimals") else None
 
             # _____attr[format]
             format_str = tag.get("format").split(":")[-1] if tag.get("format") else None
